requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
passlib==1.7.4
pyjwt==2.10.1
python-multipart==0.0.20
slowapi==0.1.9
bcrypt==3.2.0